requirements.txt
torch>=2.1.0
flash-attn>=2.5.0
deepspeed>=0.14.0
wandb>=0.16.0
transformers>=4.40.0
datasets>=2.18.0
huggingface_hub>=0.22.0
lm-eval>=0.4.0
pyyaml>=6.0
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.11.0
pandas>=2.0.0
